# Conflicts:
#	src/main/Constants/Constants.java
#	src/main/use_case/StartScreen/StartGame.java
#	src/main/view/StartScreen.java

+++ src/main/entity/tiles/PropertyTile.java
@@ -1,14 +1,13 @@
 package main.entity.tiles;
 
 import main.entity.players.rentModifier;
-import main.use_case.Tile;
 import main.entity.players.Player;
 
 /*
  * A purchasable board tile that can collect rent.
  * */
 public class PropertyTile extends Tile {
-    private final float price;
+    private final int price;
     private final float rent;
     private Player owner; //null if not owned
 
@@ -17,7 +16,7 @@ public class PropertyTile extends Tile {
      * @param price purchase price
      * @param rent  base rent amount
      */
-    public PropertyTile(String name, float price, float rent) {
+    public PropertyTile(String name, int price, float rent) {
         super(name);
         this.price = price;
         this.rent = rent;
@@ -26,7 +25,7 @@ public class PropertyTile extends Tile {
     /**
      * @return purchase price
      */
-    public float getPrice() {
+    public int getPrice() {
         return price;
     }
